+++ packages/manager/modules/sign-up/src/components/siret/siret.component.js
@@ -14,5 +14,6 @@ export default {
     onFieldError: '<?',
     mode: '<',
     fieldToFocus: '<',
+    isIndianSubsidiary: '<?',
   },
 };