+++ packages/manager/tools/sao-ovh-manager-app/template/src/app.module.js
@@ -5,19 +5,33 @@ import has from 'lodash/has';
 import isString from 'lodash/isString';
 import uiRouter, {RejectType} from '@uirouter/angularjs';
 import '@ovh-ux/ui-kit';
+import '@ovh-ux/ng-at-internet';
 import { isTopLevelApplication } from '@ovh-ux/manager-config';
-import { registerAtInternet } from '@ovh-ux/ng-shell-tracking';
 import { registerCoreModule } from '@ovh-ux/manager-core';
 import ngOvhSsoAuth from '@ovh-ux/ng-ovh-sso-auth';
 import ngUiRouterBreadcrumb from '@ovh-ux/ng-ui-router-breadcrumb';
-import <%= pascalcasedName %> from '@ovh-ux/manager-<%= name %>';
+import ovhManagerAtInternetConfiguration from '@ovh-ux/manager-at-internet-configuration';
+import { registerAtInternet } from '@ovh-ux/ng-shell-tracking';
+// TODO: Change to '@ovh-ux/manager-<%= name %>' when module is deployed
+// import <%= pascalcasedName %> from '@ovh-ux/manager-<%= name %>';
+import <%= pascalcasedName %> from '../../../modules/<%= name %>/src';
 import errorPage from './error';
 
+import TRACKING from './tracking/at-internet.constants'
 import '@ovh-ux/ui-kit/dist/css/oui.css';
 
 
 export default async (containerEl, shellClient) => {
   const moduleName = '<%= pascalcasedName %>App';
+
+  const routingConfig = /* @ngInject */($urlRouterProvider) => {
+    $urlRouterProvider.otherwise('/<%= name %>');
+  };
+
+  const trackingConfig = /* @ngInject */ (atInternetConfigurationProvider) => {
+    atInternetConfigurationProvider.setSkipInit(true);
+    atInternetConfigurationProvider.setPrefix('<%= pascalcasedName %>App');
+  };
 
   const [environment, locale] = await Promise.all([
     shellClient.environment.getEnvironment(),
@@ -30,12 +44,84 @@ export default async (containerEl, shellClient) => {
     },
   };
 
+  const ssoAuthConfig = /* @ngInject */ (ssoAuthenticationProvider) => {
+    ssoAuthenticationProvider.setOnLogin(() => {
+      shellClient.auth.login();
+    });
+    ssoAuthenticationProvider.setOnLogout(() => {
+      shellClient.auth.logout();
+    });
+  };
+
+  const calendarConfigProvider = /* @ngInject */ (ouiCalendarConfigurationProvider) => {
+    const [lang] = locale.split('_');
+    return import(`flatpickr/dist/l10n/${lang}.js`)
+      .then((module) => {
+        ouiCalendarConfigurationProvider.setLocale(module.default[lang]);
+      })
+      .catch(() => {});
+  };
+
+  const broadcastAppStarted = /* @ngInject */($rootScope, $transitions) => {
+    const unregisterHook = $transitions.onSuccess({}, async () => {
+      if (!isTopLevelApplication()) {
+        await shellClient.ux.hidePreloader();
+      }
+      $rootScope.$broadcast('app:started');
+      unregisterHook();
+    });
+  };
+
+  const transitionsConfig = /* @ngInject */ ($transitions) => {
+    if (!isTopLevelApplication()) {
+      $transitions.onBefore({}, (transition) => {
+        if (
+          !transition.ignored() &&
+          transition.from().name !== '' &&
+          transition.entering().length > 0
+        ) {
+          shellClient.ux.startProgress();
+        }
+      });
+
+      $transitions.onSuccess({}, () => {
+        shellClient.ux.stopProgress();
+      });
+
+      $transitions.onError({}, (transition) => {
+        if (!transition.error().redirected) {
+          shellClient.ux.stopProgress();
+        }
+      });
+    }
+  };
+
+  const defaultErrorHandler = /* @ngInject */($state) => {
+    $state.defaultErrorHandler((error) => {
+      if (error.type === RejectType.ERROR) {
+        $state.go(
+          'error',
+          {
+            detail: {
+              message: get(error.detail, 'data.message'),
+              code: has(error.detail, 'headers')
+                ? error.detail.headers('x-ovh-queryId')
+                : null,
+            },
+          },
+          { location: false },
+        );
+      }
+    });
+  };
+
   angular
     .module(
       moduleName,
       [
         registerCoreModule(environment, coreCallbacks),
         registerAtInternet(shellClient.tracking),
+        ovhManagerAtInternetConfiguration,
         ngOvhSsoAuth,
         ngUiRouterBreadcrumb,
         'oui',
@@ -49,91 +135,16 @@ export default async (containerEl, shellClient) => {
     .config(
       /* @ngInject */($locationProvider) => $locationProvider.hashPrefix(''),
     )
-    .config(
-      /* @ngInject */($urlRouterProvider) => {
-        $urlRouterProvider.otherwise('/<%= name %>');
-      },
-    )
-    .config(
-      /* @ngInject */ (ssoAuthenticationProvider) => {
-        ssoAuthenticationProvider.setOnLogin(() => {
-          shellClient.auth.login();
-        });
-        ssoAuthenticationProvider.setOnLogout(() => {
-          shellClient.auth.logout();
-        });
-      },
-    )
-    // @TODO initialize tracking configuration here
-    // .config(async () => {
-    //   await shellClient.tracking.setConfig(TRACKING);
-    // })
-    .config(
-      /* @ngInject */ (ouiCalendarConfigurationProvider) => {
-        const [lang] = locale.split('_');
-        return import(`flatpickr/dist/l10n/${lang}.js`)
-          .then((module) => {
-            ouiCalendarConfigurationProvider.setLocale(module.default[lang]);
-          })
-          .catch(() => {});
-      },
-    )
-    .run(
-      /* @ngInject */($rootScope, $transitions) => {
-        const unregisterHook = $transitions.onSuccess({}, async () => {
-          if (!isTopLevelApplication()) {
-            await shellClient.ux.hidePreloader();
-          }
-          $rootScope.$broadcast('app:started');
-          unregisterHook();
-        });
-      },
-    )
-    .run(
-      /* @ngInject */ ($transitions) => {
-        if (!isTopLevelApplication()) {
-          $transitions.onBefore({}, (transition) => {
-            if (
-              !transition.ignored() &&
-              transition.from().name !== '' &&
-              transition.entering().length > 0
-            ) {
-              shellClient.ux.startProgress();
-            }
-          });
-
-          $transitions.onSuccess({}, () => {
-            shellClient.ux.stopProgress();
-          });
-
-          $transitions.onError({}, (transition) => {
-            if (!transition.error().redirected) {
-              shellClient.ux.stopProgress();
-            }
-          });
-        }
-      },
-    )
-    .run(
-      /* @ngInject */($state) => {
-        $state.defaultErrorHandler((error) => {
-          if (error.type === RejectType.ERROR) {
-            $state.go(
-              'error',
-              {
-                detail: {
-                  message: get(error.detail, 'data.message'),
-                  code: has(error.detail, 'headers')
-                    ? error.detail.headers('x-ovh-queryId')
-                    : null,
-                },
-              },
-              { location: false },
-            );
-          }
-        });
-      },
-    );
+    .config(routingConfig)
+    .config(ssoAuthConfig)
+    .config(async () => {
+      await shellClient.tracking.setConfig(environment.getRegion(), TRACKING);
+    })
+    .config(trackingConfig)
+    .config(calendarConfigProvider)
+    .run(broadcastAppStarted)
+    .run(transitionsConfig)
+    .run(defaultErrorHandler);
 
   angular.bootstrap(containerEl, [moduleName], {
     strictDi: true,