+++ packages/manager/tools/sao-ovh-manager-module/lib/update-pkg.js
@@ -18,7 +18,7 @@ module.exports = ({ name, description }) => ({
     '@ovh-ux/manager-core': '^10.0.0 || ^11.0.0',
     '@ovh-ux/manager-ng-layout-helpers': '^2.0.0',
     '@ovh-ux/ng-ui-router-breadcrumb': '^1.0.0',
-    '@ovh-ux/ui-kit': '^4.4.3',
+    '@ovh-ux/ui-kit': '^6.5.1',
     '@uirouter/angularjs': '^1.0.23',
     angular: '^1.7.5',
     'angular-translate': '^2.18.1',