+++ packages/components/ng-at-internet/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.10.4](https://github.com/ovh/manager/compare/@ovh-ux/ng-at-internet@5.10.3...@ovh-ux/ng-at-internet@5.10.4) (2023-08-10)
+
+**Note:** Version bump only for package @ovh-ux/ng-at-internet
+
+
+
+
+
 ## [5.10.3](https://github.com/ovh/manager/compare/@ovh-ux/ng-at-internet@5.10.2...@ovh-ux/ng-at-internet@5.10.3) (2023-07-24)
 
 **Note:** Version bump only for package @ovh-ux/ng-at-internet