+++ packages/manager/apps/dbaas-logs/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.14](https://github.com/ovh/manager/compare/@ovh-ux/manager-dbaas-logs-app@1.10.13...@ovh-ux/manager-dbaas-logs-app@1.10.14) (2023-08-17)
+
+**Note:** Version bump only for package @ovh-ux/manager-dbaas-logs-app
+
+
+
+
+
+## [1.10.13](https://github.com/ovh/manager/compare/@ovh-ux/manager-dbaas-logs-app@1.10.12...@ovh-ux/manager-dbaas-logs-app@1.10.13) (2023-08-10)
+
+**Note:** Version bump only for package @ovh-ux/manager-dbaas-logs-app
+
+
+
+
+
 ## [1.10.12](https://github.com/ovh/manager/compare/@ovh-ux/manager-dbaas-logs-app@1.10.11...@ovh-ux/manager-dbaas-logs-app@1.10.12) (2023-08-01)
 
 **Note:** Version bump only for package @ovh-ux/manager-dbaas-logs-app