+++ packages/manager/apps/web-paas/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.26](https://github.com/ovh/manager/compare/@ovh-ux/manager-web-paas-app@1.6.25...@ovh-ux/manager-web-paas-app@1.6.26) (2023-08-17)
+
+**Note:** Version bump only for package @ovh-ux/manager-web-paas-app
+
+
+
+
+
+## [1.6.25](https://github.com/ovh/manager/compare/@ovh-ux/manager-web-paas-app@1.6.24...@ovh-ux/manager-web-paas-app@1.6.25) (2023-08-10)
+
+**Note:** Version bump only for package @ovh-ux/manager-web-paas-app
+
+
+
+
+
 ## [1.6.24](https://github.com/ovh/manager/compare/@ovh-ux/manager-web-paas-app@1.6.23...@ovh-ux/manager-web-paas-app@1.6.24) (2023-08-08)
 
 **Note:** Version bump only for package @ovh-ux/manager-web-paas-app