+++ packages/manager/modules/sign-up/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.8.0](https://github.com/ovh/manager/compare/@ovh-ux/sign-up@2.7.0...@ovh-ux/sign-up@2.8.0) (2023-08-17)
+
+
+### Features
+
+* **dedicated.account:** customer account changes ([#9728](https://github.com/ovh/manager/issues/9728)) ([bc9d697](https://github.com/ovh/manager/commit/bc9d6972a812057e19b004eea440960492671122))
+
+
+
+
+
+# [2.7.0](https://github.com/ovh/manager/compare/@ovh-ux/sign-up@2.6.4...@ovh-ux/sign-up@2.7.0) (2023-08-10)
+
+
+### Features
+
+* sms marketing consent ([#9797](https://github.com/ovh/manager/issues/9797)) ([5ca7626](https://github.com/ovh/manager/commit/5ca7626097dc31ef1b3565626ebc024cc816a150))
+
+
+
+
+
 ## [2.6.4](https://github.com/ovh/manager/compare/@ovh-ux/sign-up@2.6.3...@ovh-ux/sign-up@2.6.4) (2023-07-24)
 
 **Note:** Version bump only for package @ovh-ux/sign-up